Bump PKHeX to 24.3.10 (#246)
* Fix OutOfRange Error in sbm WriteLargeBytes

Probably a better way to do this but whatever, it works

* Update PKHeX.Core to 24.3.10

* Update version

+++ PKHeX.Core.AutoMod/AutoMod/Regeneration/RegenTemplateExtensions.cs
@@ -5,7 +5,7 @@ namespace PKHeX.Core.AutoMod
 {
     public static class RegenTemplateExtensions
     {
-        public static void SanitizeForm(this RegenTemplate set, int gen)
+        public static void SanitizeForm(this RegenTemplate set, byte gen)
         {
             if (gen is 9)
             {
